refactor(tocco-ui): update react-dropzone
Refs: TOCDEV-5019
Changelog: update react-dropzone dependancy
