+++ packages/core/tocco-ui/package.json
@@ -24,7 +24,7 @@
     "flatpickr": "4.6.9",
     "qrcode.react": "^1.0.1",
     "react-autosuggest": "^10.1.0",
-    "react-dropzone": "^4.1.3",
+    "react-dropzone": "^12.0.4",
     "react-number-format": "^4.9.1",
     "react-quill": "^1.3.5",
     "react-select": "^5.2.2",
