+++ packages/core/tocco-ui/src/Upload/UploadInput.js
@@ -1,11 +1,19 @@
 import PropTypes from 'prop-types'
 import React from 'react'
+import Dropzone from 'react-dropzone'
 
 import Button from '../Button'
-import {StyledUploadInput, StyledDropzone} from './StyledUploadInput'
+import {StyledUploadInput, StyledSection, StyledDropzone} from './StyledComponents'
 
 const UploadInput = ({onDrop: onDropProp, immutable, text, onChoose}) => {
-  const onDrop = files => onDropProp(files[0])
+  const onDrop = acceptedFiles => {
+    const previewAddedFiles = acceptedFiles.map(file =>
+      Object.assign(file, {
+        preview: URL.createObjectURL(file)
+      })
+    )
+    onDropProp(previewAddedFiles[0])
+  }
   const handleClick = e => {
     onChoose()
     e.stopPropagation()
@@ -13,9 +21,16 @@ const UploadInput = ({onDrop: onDropProp, immutable, text, onChoose}) => {
 
   return (
     <StyledUploadInput immutable={immutable}>
-      <StyledDropzone disabled={immutable} multiple={false} onDrop={onDrop} title={text || 'drag and drop or click'}>
-        <Button icon="arrow-to-top" />
-      </StyledDropzone>
+      <Dropzone onDrop={onDrop} multiple={false} disabled={immutable} title={text || 'drag and drop or click'}>
+        {({getRootProps, getInputProps}) => (
+          <StyledSection>
+            <StyledDropzone {...getRootProps()}>
+              <input {...getInputProps()} />
+              <Button icon="arrow-to-top" />
+            </StyledDropzone>
+          </StyledSection>
+        )}
+      </Dropzone>
       {onChoose && <Button onClick={handleClick} icon="folder" label={'browse docs'} iconOnly={true} />}
     </StyledUploadInput>
   )
