+++ packages/core/tocco-ui/src/Upload/index.js
@@ -1,5 +1,4 @@
-import {StyledUploadInput} from './StyledUploadInput'
-import {StyledView} from './StyledView'
+import {StyledUploadInput, StyledView} from './StyledComponents'
 import Upload from './Upload'
 
 export {StyledUploadInput, StyledView, Upload as default}
